+++ src/app/components/modals/modal-delete-generic/modal-delete-generic.component.scss
