+++ src/app/components/modals/modal-highlight-delete/modal-highlight-delete.component.html
@@ -1,0 +1,62 @@
+<!--
+
+  Digital Excellence Copyright (C) 2020 Brend Smits
+
+   This program is free software: you can redistribute it and/or modify
+   it under the terms of the GNU Lesser General Public License as published
+   by the Free Software Foundation version 3 of the License.
+
+   This program is distributed in the hope that it will be useful,
+   but WITHOUT ANY WARRANTY; without even the implied warranty
+   of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
+   See the GNU Lesser General Public License for more details.
+
+   You can find a copy of the GNU Lesser General Public License
+   along with this program, in the LICENSE.md file in the root project directory.
+   If not, see https://www.gnu.org/licenses/lgpl-3.0.txt
+
+ -->
+
+<div class="modal-delete-container">
+  <button type="button" class="close pull-right" aria-label="Close" (click)="bsModalRef.hide()">
+    <span aria-hidden="true">&times;</span>
+  </button>
+  <div class="modal-body">
+    <h2>Delete highlight</h2>
+    <div class="row">
+      <div class="col-12">
+        <div class="highlight-form__checkbox">
+          <form [formGroup]="deleteHighlightForm" class="highlight-form">
+            <div *ngFor="let highlight of highlights">
+              <div *ngIf="highlight.isNeverEnding; else displayDateTimes">
+                <div class="highlight-wrapper">
+                  <input type="checkbox" id="neverEndingCheckbox" [formControlName]="highlight.id" class="" />
+                  <label for="neverEndingCheckbox"><strong>Never Ending</strong></label>
+
+                </div>
+              </div>
+              <ng-template #displayDateTimes>
+                <div class="highlight-wrapper">
+                  <input type="checkbox" id="checkbox" [formControlName]="highlight.id" class="" />
+                  <label for="checkbox">
+                    <div class="row no-gutters">
+                      <div class="col-6"><strong>Start date:</strong>{{highlight.startDate}}</div>
+                      <div class="col-6"><strong>End date:</strong>{{highlight.endDate}}</div>
+                    </div>
+                  </label>
+
+                </div>
+              </ng-template>
+            </div>
+          </form>
+        </div>
+      </div>
+    </div>
+  </div>
+  <div class="row">
+    <div class="col-12 modal-buttons">
+      <button type="button" class="cancel" (click)="onClickDeny()">Cancel</button>
+      <button type="button" class="primary" (click)="onClickConfirm()">Delete</button>
+    </div>
+  </div>
+</div>
